Revert "#17593 Add Extension Points For Pre and Post Collection of Scores in …"
This reverts commit eda3a48f69c46915c166b2573cfb9f1ce4ff52b1.
